file-operations: Do not report remaining time if progress is not reported
Some GVfs backends doesn't report progress except the total file size. With
the previous patch applied, the remaining time is calculated after each
successfully transfered file for such backends. This is fine for small files,
but looks weird for big files as the estimated time is not updated regularly,
so it looks like hang. One possible solution would be to update the remaining
time periodically even if progress is not reported from GIO. However, this
could again lead to situations, which would look like that the transfer is
hanged. Let's do not report remaining time at all when progress is not reported
to reduce confusion.